Add multiple saves / change event proportions (#149)
* feat: Add multiple rows per block / profile

Adds the ability to fake how many times a course was published or a user
profile saved. In the real world these are drivers of table size and we
should exercise the "most recent block" / "most recent profile" code
more.

* refactor: Don't enforce backend type on S3 load

When doing S3 CSV generation and loads it required a lot of unnecessary
switching back and forth between types in the config. Now as long as the
correct keys exist it won't complain.

* refactor: Change event proportions

Based on feedback from live sites, we've recalculated the proportions of
event types. This takes an average of 4 different live sites to find a
reasonable middle ground. Site configuration and course design play
large roles in these distributions in real life.